Move experimental.kernel_iface into top-level module.
