+++ numba_dpex/kernel_iface/__init__.py
@@ -2,7 +2,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""The kernel_iface provides a set of Python classes and functions that are
+"""
+The kernel_iface provides a set of Python classes and functions that are
 analogous to the C++ SYCL API. The kernel_iface API is meant to allow
 prototyping SYCL-like kernels in pure Python before compiling them using
 numba_dpex.kernel.
