+++ numba_dpex/tests/experimental/kernel_iface/test_memory_enum_compilation.py
@@ -6,11 +6,7 @@ import dpnp
 
 import numba_dpex.experimental as exp_dpex
 from numba_dpex import Range
-from numba_dpex.experimental.kernel_iface import (
-    AddressSpace,
-    MemoryOrder,
-    MemoryScope,
-)
+from numba_dpex.kernel_iface import AddressSpace, MemoryOrder, MemoryScope
 
 
 def test_compilation_of_memory_order():
